Adjusted Sharpe ratio of the return distribution

Adjusted Sharpe ratio was introduced by Pezier and White (2006) to adjusts for skewness and kurtosis by incorporating a penalty factor for negative skewness and excess kurtosis.

AdjustedSharpeRatio(R, Rf = 0, ...)

Arguments

R
an xts, vector, matrix, data frame, timeSeries or zoo object of asset returns

Rf
the risk free rate

…
any other passthru parameters

Details

$$Adjusted Sharpe Ratio = SR * [1 + (\frac{S}{6}) * SR - (\frac{K - 3}{24}) * SR^2]$$

where \(SR\) is the sharpe ratio with data annualized, \(S\) is the skewness and \(K\) is the kurtosis

References

Carl Bacon, Practical portfolio performance measurement and attribution, second edition 2008 p.99

Pezier, Jaques and White, Anthony. 2006. The Relative Merits of Investable Hedge Fund Indices and of Funds of Hedge Funds in Optimal Passive Portfolios. http://econpapers.repec.org/paper/rdgicmadp/icma-dp2006-10.htm

See also

SharpeRatio.annualized

Examples

data(portfolio_bacon)
print(AdjustedSharpeRatio(portfolio_bacon[,1])) #expected 0.7591435

#>                                 portfolio.monthly.return....
#> Annualized Sharpe Ratio (Rf=0%)                    0.7591435

data(managers)
print(AdjustedSharpeRatio(managers['1996']))

#>                                           HAM1    HAM2      HAM3     HAM4 HAM5
#> Adjusted Sharpe ratio (Risk free = 0) 2.045968 14.5593 0.9322736 1.883368   NA
#>                                       HAM6 EDHEC LS EQ SP500 TR   US 10Y TR
#> Adjusted Sharpe ratio (Risk free = 0)   NA          NA 1.986962 0.006312774
#>                                        US 3m TR
#> Adjusted Sharpe ratio (Risk free = 0) -576.9696