Appraisal ratio of the return distribution

Appraisal ratio is the Jensen's alpha adjusted for specific risk. The numerator is divided by specific risk instead of total risk.

AppraisalRatio(Ra, Rb, Rf = 0, method = c("appraisal", "modified",
  "alternative"), ...)

Arguments

Ra
an xts, vector, matrix, data frame, timeSeries or zoo object of asset returns

Rb
return vector of the benchmark asset

Rf
risk free rate, in same period as your returns

method
is one of "appraisal" to calculate appraisal ratio, "modified" to calculate modified Jensen's alpha or "alternative" to calculate alternative Jensen's alpha.

…
any other passthru parameters

Details

Modified Jensen's alpha is Jensen's alpha divided by beta.

Alternative Jensen's alpha is Jensen's alpha divided by systematic risk. $$Appraisal ratio = \frac{\alpha}{\sigma_{\epsilon}}$$ $$Modified Jensen's alpha = \frac{\alpha}{\beta}$$ $$Alternative Jensen's alpha = \frac{\alpha}{\sigma_S}$$

where \(alpha\) is the Jensen's alpha, \(\sigma_{epsilon}\) is the specific risk, \(\sigma_S\) is the systematic risk.

References

Carl Bacon, Practical portfolio performance measurement and attribution, second edition 2008 p.77

Examples

data(portfolio_bacon)
print(AppraisalRatio(portfolio_bacon[,1], portfolio_bacon[,2], method="appraisal")) #expected -0.430

#> [1] -0.4302756

print(AppraisalRatio(portfolio_bacon[,1], portfolio_bacon[,2], method="modified"))

#> [1] -0.01418576

print(AppraisalRatio(portfolio_bacon[,1], portfolio_bacon[,2], method="alternative"))

#> [1] -0.1066928

data(managers)
print(AppraisalRatio(managers['1996',1], managers['1996',8]))

#> [1] 1.623025

print(AppraisalRatio(managers['1996',1:5], managers['1996',8]))

#>                                     HAM1 HAM2     HAM3      HAM4 HAM5
#> Appraisal ratio (Risk free = 0) 1.623025   NA 3.527723 0.7070483   NA