Functions to calculate systematic or beta co-moments of return series

calculate higher co-moment betas, or 'systematic' variance, skewness, and kurtosis

BetaCoVariance(Ra, Rb)

BetaCoSkewness(Ra, Rb, test = FALSE)

BetaCoKurtosis(Ra, Rb)

Arguments

Ra
an xts, vector, matrix, data frame, timeSeries or zoo object of asset returns

Rb
an xts, vector, matrix, data frame, timeSeries or zoo object of index, benchmark, or secondary asset returns to compare against

test
condition not implemented yet

References

Boudt, Kris, Brian G. Peterson, and Christophe Croux. 2008. Estimation and Decomposition of Downside Risk for Portfolios with Non-Normal Returns. Journal of Risk. Winter.

Martellini, Lionel, and Volker Ziemann. 2007. Improved Forecasts of Higher-Order Comoments and Implications for Portfolio Selection. EDHEC Risk and Asset Management Research Centre working paper.

See also

CoMoments

Examples

data(managers)

BetaCoVariance(managers[, "HAM2", drop=FALSE], managers[, "SP500 TR", drop=FALSE])

#> [1] 0.3431621

BetaCoSkewness(managers[, "HAM2", drop=FALSE], managers[, "SP500 TR", drop=FALSE])

#> [1] 0.04542927

BetaCoKurtosis(managers[, "HAM2", drop=FALSE], managers[, "SP500 TR", drop=FALSE])

#> [1] 0.1988373

BetaCoKurtosis(managers[,1:6], managers[,8,drop=FALSE])

#>                                HAM1      HAM2     HAM3      HAM4      HAM5
#> Beta Cokurtosis: SP500 TR 0.4814681 0.1988373 0.506819 0.8483555 0.2738611
#>                                HAM6
#> Beta Cokurtosis: SP500 TR 0.1541281

BetaCoKurtosis(managers[,1:6], managers[,8:7])

#>                                   HAM1      HAM2     HAM3      HAM4      HAM5
#> Beta Cokurtosis: SP500 TR    0.4814681 0.1988373 0.506819 0.8483555 0.2738611
#> Beta Cokurtosis: EDHEC LS EQ 0.7100547 1.2676023 1.426660 1.4533001 1.2831205
#>                                   HAM6
#> Beta Cokurtosis: SP500 TR    0.1541281
#> Beta Cokurtosis: EDHEC LS EQ 0.8618328