calculate single factor model (CAPM) alpha

This is a wrapper for calculating a single factor model (CAPM) alpha.

CAPM.alpha(Ra, Rb, Rf = 0)

Arguments

Ra
an xts, vector, matrix, data frame, timeSeries or zoo object of asset returns

Rb
return vector of the benchmark asset

Rf
risk free rate, in same period as your returns

Details

"Alpha" purports to be a measure of a manager's skill by measuring the portion of the managers returns that are not attributable to "Beta", or the portion of performance attributable to a benchmark.

While the classical CAPM has been almost completely discredited by the literature, it is an example of a simple single factor model, comparing an asset to any arbitrary benchmark.

References

Sharpe, W.F. Capital Asset Prices: A theory of market equilibrium under conditions of risk. Journal of finance, vol 19, 1964, 425-442. Ruppert, David. Statistics and Finance, an Introduction. Springer. 2004.

See also

CAPM.beta CAPM.utils

Examples

# First we load the data
    data(managers)
    CAPM.alpha(managers[,1,drop=FALSE],
                        managers[,8,drop=FALSE],
                        Rf=.035/12)

#> [1] 0.005960609

    CAPM.alpha(managers[,1,drop=FALSE],
                        managers[,8,drop=FALSE],
                        Rf = managers[,10,drop=FALSE])

#> [1] 0.005774729

    CAPM.alpha(managers[,1:6],
                        managers[,8,drop=FALSE],
                        Rf=.035/12)

#>                        HAM1        HAM2        HAM3        HAM4        HAM5
#> Alpha: SP500 TR 0.005960609 0.009232784 0.006327417 0.004144392 0.001425414
#>                        HAM6
#> Alpha: SP500 TR 0.007244216

    CAPM.alpha(managers[,1:6],
                        managers[,8,drop=FALSE],
                        Rf = managers[,10,drop=FALSE])

#>                        HAM1        HAM2        HAM3        HAM4        HAM5
#> Alpha: SP500 TR 0.005774729 0.009092773 0.006216498 0.004029731 0.001733199
#>                        HAM6
#> Alpha: SP500 TR 0.007837454

    CAPM.alpha(managers[,1:6],
                        managers[,8:7,drop=FALSE],
                        Rf=.035/12)

#>                           HAM1        HAM2         HAM3         HAM4
#> Alpha: SP500 TR    0.005960609 0.009232784  0.006327417  0.004144392
#> Alpha: EDHEC LS EQ 0.003190734 0.001627709 -0.001716897 -0.003357739
#>                            HAM5        HAM6
#> Alpha: SP500 TR     0.001425414 0.007244216
#> Alpha: EDHEC LS EQ -0.002185688 0.004197041

    CAPM.alpha(managers[,1:6],
                        managers[,8:7,drop=FALSE],
                        Rf = managers[,10,drop=FALSE])

#>                           HAM1        HAM2         HAM3         HAM4
#> Alpha: SP500 TR    0.005774729 0.009092773  0.006216498  0.004029731
#> Alpha: EDHEC LS EQ 0.003134750 0.001732335 -0.001563904 -0.003406683
#>                            HAM5        HAM6
#> Alpha: SP500 TR     0.001733199 0.007837454
#> Alpha: EDHEC LS EQ -0.002255793 0.004212371