Calculate Uryasev's proposed Conditional Drawdown at Risk (CDD or CDaR) measure

For some confidence level \(p\), the conditional drawdown is the the mean of the worst \(p\%\) drawdowns.

CDD(R, weights = NULL, geometric = TRUE, invert = TRUE, p = 0.95, ...)

Arguments

R
an xts, vector, matrix, data frame, timeSeries or zoo object of asset returns

weights
portfolio weighting vector, default NULL, see Details

geometric
utilize geometric chaining (TRUE) or simple/arithmetic chaining (FALSE) to aggregate returns, default TRUE

invert
TRUE/FALSE whether to invert the drawdown measure. see Details.

p
confidence level for calculation, default p=0.95

…
any other passthru parameters

References

Chekhlov, A., Uryasev, S., and M. Zabarankin. Portfolio Optimization With Drawdown Constraints. B. Scherer (Ed.) Asset and Liability Management Tools, Risk Books, London, 2003 http://www.ise.ufl.edu/uryasev/drawdown.pdf

See also

ES maxDrawdown

Examples

data(edhec)
t(round(CDD(edhec),4))

#>                        Conditional Drawdown 5%
#> Convertible Arbitrage                   0.0927
#> CTA Global                              0.0548
#> Distressed Securities                   0.0695
#> Emerging Markets                        0.3433
#> Equity Market Neutral                   0.0111
#> Event Driven                            0.0813
#> Fixed Income Arbitrage                  0.1086
#> Global Macro                            0.0407
#> Long/Short Equity                       0.0842
#> Merger Arbitrage                        0.0527
#> Relative Value                          0.0439
#> Short Selling                           0.4293
#> Funds of Funds                          0.0563