calculate a Calmar or Sterling reward/risk ratio Calmar and Sterling Ratios are yet another method of creating a risk-adjusted measure for ranking investments similar to the SharpeRatio.

Both the Calmar and the Sterling ratio are the ratio of annualized return over the absolute value of the maximum drawdown of an investment. The Sterling ratio adds an excess risk measure to the maximum drawdown, traditionally and defaulting to 10%.

CalmarRatio(R, scale = NA)

SterlingRatio(R, scale = NA, excess = 0.1)

Arguments

R
an xts, vector, matrix, data frame, timeSeries or zoo object of asset returns

scale
number of periods in a year (daily scale = 252, monthly scale = 12, quarterly scale = 4)

excess
for Sterling Ratio, excess amount to add to the max drawdown, traditionally and default .1 (10%)

Details

It is also traditional to use a three year return series for these calculations, although the functions included here make no effort to determine the length of your series. If you want to use a subset of your series, you'll need to truncate or subset the input data to the desired length.

Many other measures have been proposed to do similar reward to risk ranking. It is the opinion of this author that newer measures such as Sortino's UpsidePotentialRatio or Favre's modified SharpeRatio are both “better” measures, and should be preferred to the Calmar or Sterling Ratio.

References

Bacon, Carl. Practical Portfolio Performance Measurement and Attribution. Wiley. 2004.

See also

Return.annualized, maxDrawdown, SharpeRatio.modified, UpsidePotentialRatio

Examples

    data(managers)
    CalmarRatio(managers[,1,drop=FALSE])

#>                   HAM1
#> Calmar Ratio 0.9061697

    CalmarRatio(managers[,1:6])

#>                   HAM1     HAM2      HAM3      HAM4      HAM5     HAM6
#> Calmar Ratio 0.9061697 0.728094 0.5225829 0.4227315 0.1095909 1.742525

    SterlingRatio(managers[,1,drop=FALSE])

#>                                    HAM1
#> Sterling Ratio (Excess = 10%) 0.5462542

    SterlingRatio(managers[,1:6])

#>                                    HAM1      HAM2      HAM3      HAM4
#> Sterling Ratio (Excess = 10%) 0.5462542 0.5138746 0.3883671 0.3136025
#>                                     HAM5      HAM6
#> Sterling Ratio (Excess = 10%) 0.08471255 0.7678475