Kappa of the return distribution

Introduced by Kaplan and Knowles (2004), Kappa is a generalized downside risk-adjusted performance measure.

Kappa(R, MAR, l, ...)

Arguments

R
an xts, vector, matrix, data frame, timeSeries or zoo object of asset returns

MAR
Minimum Acceptable Return, in the same periodicity as your returns

l
the coefficient of the Kappa

…
any other passthru parameters

Details

To calculate it, we take the difference of the mean of the distribution to the target and we divide it by the l-root of the lth lower partial moment. To calculate the lth lower partial moment we take the subset of returns below the target and we sum the differences of the target to these returns. We then return return this sum divided by the length of the whole distribution. $$Kappa(R, MAR, l) = \frac{r_{p}-MAR}{\sqrt[l]{\frac{1}{n}*\sum^n_{t=1} max(MAR-R_{t}, 0)^l}}$$

For l=1 kappa is the Sharpe-omega ratio and for l=2 kappa is the sortino ratio.

Kappa should only be used to rank portfolios as it is difficult to interpret the absolute differences between kappas. The higher the kappa is, the better.

References

Carl Bacon, Practical portfolio performance measurement and attribution, second edition 2008 p.96

Examples

l = 2

data(portfolio_bacon)
MAR = 0.005
print(Kappa(portfolio_bacon[,1], MAR, l)) #expected 0.157

#> [1] 0.1566371

data(managers)
MAR = 0
print(Kappa(managers['1996'], MAR, l))

#>                      HAM1     HAM2     HAM3    HAM4 HAM5 HAM6 EDHEC LS EQ
#> kappa (MAR = 0%) 1.492063 1061.685 2.235197 1.14188  NaN  NaN         NaN
#>                  SP500 TR  US 10Y TR US 3m TR
#> kappa (MAR = 0%) 1.274332 0.01674457      Inf

print(Kappa(managers['1996',1], MAR, l)) #expected 1.493

#> [1] 1.492063