Market timing models

Allows to estimate Treynor-Mazuy or Merton-Henriksson market timing model. The Treynor-Mazuy model is essentially a quadratic extension of the basic CAPM. It is estimated using a multiple regression. The second term in the regression is the value of excess return squared. If the gamma coefficient in the regression is positive, then the estimated equation describes a convex upward-sloping regression "line". The quadratic regression is: $$R_{p}-R_{f}=\alpha+\beta (R_{b} - R_{f})+\gamma (R_{b}-R_{f})^2+ \varepsilon_{p}$$ \(\gamma\) is a measure of the curvature of the regression line. If \(\gamma\) is positive, this would indicate that the manager's investment strategy demonstrates market timing ability.

MarketTiming(Ra, Rb, Rf = 0, method = c("TM", "HM"))

Arguments

Ra
an xts, vector, matrix, data frame, timeSeries or zoo object of the asset returns

Rb
an xts, vector, matrix, data frame, timeSeries or zoo object of the benchmark asset return

Rf
risk free rate, in same period as your returns

method
used to select between Treynor-Mazuy and Henriksson-Merton models. May be any of:

TM - Treynor-Mazuy model,

HM - Henriksson-Merton model

By default Treynor-Mazuy is selected

…
any other passthrough parameters

Details

The basic idea of the Merton-Henriksson test is to perform a multiple regression in which the dependent variable (portfolio excess return and a second variable that mimics the payoff to an option). This second variable is zero when the market excess return is at or below zero and is 1 when it is above zero: $$R_{p}-R_{f}=\alpha+\beta (R_{b}-R_{f})+\gamma D+\varepsilon_{p}$$ where all variables are familiar from the CAPM model, except for the up-market return \(D=max(0,R_{b}-R_{f})\) and market timing abilities \(\gamma\)

References

J. Christopherson, D. Carino, W. Ferson. Portfolio Performance Measurement and Benchmarking. 2009. McGraw-Hill, p. 127-133. J. L. Treynor and K. Mazuy, "Can Mutual Funds Outguess the Market?" Harvard Business Review, vol44, 1966, pp. 131-136 Roy D. Henriksson and Robert C. Merton, "On Market Timing and Investment Performance. II. Statistical Procedures for Evaluating Forecast Skills," Journal of Business, vol.54, October 1981, pp.513-533

See also

CAPM.beta

Examples

data(managers)
MarketTiming(managers[,1], managers[,8], Rf=.035/12, method = "HM")

#>                        Alpha      Beta      Gamma
#> HAM1 to SP500 TR 0.008275839 0.4555824 -0.1344417

MarketTiming(managers[80:120,1:6], managers[80:120,7], managers[80:120,10])

#>                             Alpha      Beta     Gamma
#> HAM1 to EDHEC LS EQ -0.0005755802 1.3121058 -0.405150
#> HAM2 to EDHEC LS EQ -0.0003616789 0.4370998  8.520620
#> HAM3 to EDHEC LS EQ -0.0058148518 1.1898242 11.913786
#> HAM4 to EDHEC LS EQ -0.0055113742 2.0616524 18.797340
#> HAM5 to EDHEC LS EQ  0.0005125284 1.0703704 -5.077881
#> HAM6 to EDHEC LS EQ  0.0003590925 1.2711094 -7.443428

MarketTiming(managers[80:120,1:6], managers[80:120,8:7], managers[80:120,10], method = "TM")

#>                             Alpha      Beta      Gamma
#> HAM1 to SP500 TR     0.0048833318 0.5970167 -0.2801650
#> HAM2 to SP500 TR     0.0050694247 0.1190405 -0.5000263
#> HAM3 to SP500 TR     0.0032110848 0.5272982 -0.6645684
#> HAM4 to SP500 TR     0.0094634771 0.8779523 -0.8155100
#> HAM5 to SP500 TR     0.0087234498 0.2869943 -2.7728051
#> HAM6 to SP500 TR     0.0048031173 0.2902262  0.6910898
#> HAM1 to EDHEC LS EQ -0.0005755802 1.3121058 -0.4051500
#> HAM2 to EDHEC LS EQ -0.0003616789 0.4370998  8.5206196
#> HAM3 to EDHEC LS EQ -0.0058148518 1.1898242 11.9137857
#> HAM4 to EDHEC LS EQ -0.0055113742 2.0616524 18.7973395
#> HAM5 to EDHEC LS EQ  0.0005125284 1.0703704 -5.0778814
#> HAM6 to EDHEC LS EQ  0.0003590925 1.2711094 -7.4434281