Omega-Sharpe ratio of the return distribution

The Omega-Sharpe ratio is a conversion of the omega ratio to a ranking statistic in familiar form to the Sharpe ratio.

OmegaSharpeRatio(R, MAR = 0, ...)

Arguments

R
an xts, vector, matrix, data frame, timeSeries or zoo object of asset returns

MAR
Minimum Acceptable Return, in the same periodicity as your returns

…
any other passthru parameters

Details

To calculate the Omega-Sharpe ration we subtract the target (or Minimum Acceptable Returns (MAR)) return from the portfolio return and we divide it by the opposite of the Downside Deviation. $$OmegaSharpeRatio(R,MAR) = \frac{r_p - r_t}{\sum^n_{t=1}\frac{max(r_t - r_i, 0)}{n}}$$

where \(n\) is the number of observations of the entire series

References

Carl Bacon, Practical portfolio performance measurement and attribution, second edition 2008, p.95

Examples

data(portfolio_bacon)
MAR = 0.005
print(OmegaSharpeRatio(portfolio_bacon[,1], MAR)) #expected 0.29

#> [1] 0.2917933

MAR = 0
data(managers)
print(OmegaSharpeRatio(managers['1996'], MAR))

#>                                 HAM1 HAM2     HAM3     HAM4 HAM5 HAM6
#> OmegaSharpeRatio (MAR = 0%) 3.598338 2374 5.482813 2.615074   NA   NA
#>                             EDHEC LS EQ SP500 TR  US 10Y TR US 3m TR
#> OmegaSharpeRatio (MAR = 0%)          NA 3.340625 0.02827709      Inf

print(OmegaSharpeRatio(managers['1996',1], MAR)) #expected 3.60

#> [1] 3.598338