calculate Geltner liquidity-adjusted return series

David Geltner developed a method to remove estimating or liquidity bias in real estate index returns. It has since been applied with success to other return series that show autocorrelation or illiquidity effects.

Return.Geltner(Ra, ...)

Arguments

Ra
an xts, vector, matrix, data frame, timeSeries or zoo object of asset returns

…
any other passthru parameters

Details

The theory is that by correcting for autocorrelation, you are uncovering a "true" return from a series of observed returns that contain illiquidity or manual pricing effects.

The Geltner autocorrelation adjusted return series may be calculated via: $$ $$$$ R_{G}=\frac{R_{t}-(R_{t-1}\cdot\rho_{1})}{1-\rho_{1}} $$

where \(\rho_{1}\) is the first-order autocorrelation of the return series \(R_{a}\) and \(R_{t}\) is the return of \(R_{a}\) at time \(t\) and \(R_{t-1}\) is the one-period lagged return.

References

"Edhec Funds of Hedge Funds Reporting Survey : A Return-Based Approach to Funds of Hedge Funds Reporting",Edhec Risk and Asset Management Research Centre, January 2005,p. 27

Geltner, David, 1991, Smoothing in Appraisal-Based Returns, Journal of Real Estate Finance and Economics, Vol.4, p.327-345.

Geltner, David, 1993, Estimating Market Values from Appraised Values without Assuming an Efficient Market, Journal of Real Estate Research, Vol.8, p.325-345.

Examples

data(managers)
head(Return.Geltner(managers[,1:3]),n=20)

#>                    HAM1          HAM2         HAM3
#> 1996-01-31           NA            NA           NA
#> 1996-02-29  0.022073985            NA  0.035101421
#> 1996-03-31  0.014614190            NA  0.025733921
#> 1996-04-30 -0.014834456            NA  0.045035710
#> 1996-05-31  0.011492903            NA  0.035231790
#> 1996-06-30 -0.006580742            NA -0.030766102
#> 1996-07-31 -0.027575673            NA -0.033724158
#> 1996-08-31  0.054092559            NA  0.046666996
#> 1996-09-30  0.008918922  0.1248798328  0.065436420
#> 1996-10-31  0.032086822  0.0174616062  0.039316685
#> 1996-11-30  0.012522975  0.0835177999  0.066792551
#> 1996-12-31  0.018066216  0.0189979595  0.021078844
#> 1997-01-31  0.022039189  0.0916045833  0.077495761
#> 1997-02-28 -0.002229051 -0.0297548689 -0.038213548
#> 1997-03-31  0.011078377 -0.0315013248 -0.033573000
#> 1997-04-30  0.013345946 -0.0009819489  0.029041945
#> 1997-05-31  0.051072969  0.0686636088  0.076236077
#> 1997-06-30  0.018274665  0.0555198782  0.004899082
#> 1997-07-31  0.013605069  0.1297143968  0.108829706
#> 1997-08-31  0.025634796 -0.0528443018 -0.003587969