calculate the relative return of one asset to another

Calculates the ratio of the cumulative performance for two assets through time.

Return.relative(Ra, Rb, ...)

Arguments

Ra
an xts, vector, matrix, data frame, timeSeries or zoo object of asset returns

Rb
return object for the benchmark asset

…
ignored

Value

xts or other time series of relative return

See also

chart.RelativePerformance

Examples

data(managers)
head(Return.relative(managers[,1:3], managers[,8,drop=FALSE]),n=20)

#>            HAM1/SP500 TR HAM2/SP500 TR HAM3/SP500 TR
#> 1996-01-31     0.9742747            NA      1.000870
#> 1996-02-29     0.9839276            NA      1.026455
#> 1996-03-31     0.9896776            NA      1.042925
#> 1996-04-30     0.9664645            NA      1.073965
#> 1996-05-31     0.9493173            NA      1.083912
#> 1996-06-30     0.9420352            NA      1.047090
#> 1996-07-31     0.9628313            NA      1.058593
#> 1996-08-31     0.9801813     0.9792381      1.084511
#> 1996-09-30     0.9415791     1.0199354      1.093751
#> 1996-10-31     0.9426786     1.0260891      1.106417
#> 1996-11-30     0.8900934     1.0242766      1.097159
#> 1996-12-31     0.9240553     1.0761069      1.143275
#> 1997-01-31     0.8881367     1.0932234      1.158985
#> 1997-02-28     0.8832017     1.0758672      1.107005
#> 1997-03-31     0.9297150     1.0917993      1.115663
#> 1997-04-30     0.8883924     1.0240062      1.082921
#> 1997-05-31     0.8740730     1.0172496      1.098232
#> 1997-06-30     0.8559189     1.0273754      1.056817
#> 1997-07-31     0.8050204     1.0610629      1.084716
#> 1997-08-31     0.8729867     1.1018644      1.145846