calculate a traditional or modified Sharpe Ratio of Return over StdDev or VaR or ES

The Sharpe ratio is simply the return per unit of risk (represented by variability). In the classic case, the unit of risk is the standard deviation of the returns.

SharpeRatio(R, Rf = 0, p = 0.95, FUN = c("StdDev", "VaR", "ES"),
  weights = NULL, annualize = FALSE, ...)

SharpeRatio.modified(R, Rf = 0, p = 0.95, FUN = c("StdDev", "VaR", "ES"),
  weights = NULL, ...)

Arguments

R
an xts, vector, matrix, data frame, timeSeries or zoo object of asset returns

Rf
risk free rate, in same period as your returns

p
confidence level for calculation, default p=.95

FUN
one of "StdDev" or "VaR" or "ES" to use as the denominator

weights
portfolio weighting vector, default NULL, see Details in VaR

annualize
if TRUE, annualize the measure, default FALSE

…
any other passthru parameters to the VaR or ES functions

Details

$$\frac{\overline{(R_{a}-R_{f})}}{\sqrt{\sigma_{(R_{a}-R_{f})}}}$$

William Sharpe now recommends InformationRatio preferentially to the original Sharpe Ratio.

The higher the Sharpe ratio, the better the combined performance of "risk" and return.

As noted, the traditional Sharpe Ratio is a risk-adjusted measure of return that uses standard deviation to represent risk.

A number of papers now recommend using a "modified Sharpe" ratio using a Modified Cornish-Fisher VaR or CVaR/Expected Shortfall as the measure of Risk.

We have recently extended this concept to create multivariate modified Sharpe-like Ratios for standard deviation, Gaussian VaR, modified VaR, Gaussian Expected Shortfall, and modified Expected Shortfall. See VaR and ES. You can pass additional arguments to VaR and ES via … The most important is probably the 'method' argument/

This function returns a traditional or modified Sharpe ratio for the same periodicity of the data being input (e.g., monthly data -> monthly SR)

References

Sharpe, W.F. The Sharpe Ratio,Journal of Portfolio Management,Fall 1994, 49-58.

Laurent Favre and Jose-Antonio Galeano. Mean-Modified Value-at-Risk Optimization with Hedge Funds. Journal of Alternative Investment, Fall 2002, v 5.

See also

SharpeRatio.annualized InformationRatio TrackingError ActivePremium SortinoRatio VaR ES

Examples

data(managers)
SharpeRatio(managers[,1,drop=FALSE], Rf=.035/12, FUN="StdDev")

#>                                      HAM1
#> StdDev Sharpe (Rf=0.3%, p=95%): 0.3201889

SharpeRatio(managers[,1,drop=FALSE], Rf = managers[,10,drop=FALSE], FUN="StdDev")

#>                                     HAM1
#> StdDev Sharpe (Rf=0.3%, p=95%): 0.308102

SharpeRatio(managers[,1:6], Rf=.035/12, FUN="StdDev")

#>                                      HAM1      HAM2      HAM3      HAM4
#> StdDev Sharpe (Rf=0.3%, p=95%): 0.3201889 0.3057649 0.2610141 0.1522615
#>                                       HAM5      HAM6
#> StdDev Sharpe (Rf=0.3%, p=95%): 0.02562009 0.3417545

SharpeRatio(managers[,1:6], Rf = managers[,10,drop=FALSE], FUN="StdDev")

#>                                     HAM1      HAM2      HAM3      HAM4
#> StdDev Sharpe (Rf=0.3%, p=95%): 0.308102 0.2988608 0.2525301 0.1464385
#>                                      HAM5      HAM6
#> StdDev Sharpe (Rf=0.3%, p=95%): 0.0354554 0.3785371

data(edhec)
SharpeRatio(edhec[, 6, drop = FALSE], FUN="VaR")

#>                            Event Driven
#> VaR Sharpe (Rf=0%, p=95%):    0.2710161

SharpeRatio(edhec[, 6, drop = FALSE], Rf = .04/12, FUN="VaR")

#>                              Event Driven
#> VaR Sharpe (Rf=0.3%, p=95%):    0.1524982

SharpeRatio(edhec[, 6, drop = FALSE], Rf = .04/12, FUN="VaR" , method="gaussian")

#>                              Event Driven
#> VaR Sharpe (Rf=0.3%, p=95%):    0.1909461

SharpeRatio(edhec[, 6, drop = FALSE], FUN="ES")

#>                           Event Driven
#> ES Sharpe (Rf=0%, p=95%):    0.1289815

# and all the methods
SharpeRatio(managers[,1:9], Rf = managers[,10,drop=FALSE])

#>                                      HAM1      HAM2      HAM3       HAM4
#> StdDev Sharpe (Rf=0.3%, p=95%): 0.3081020 0.2988608 0.2525301 0.14643845
#> VaR Sharpe (Rf=0.3%, p=95%):    0.2306863 0.3970699 0.2504936 0.09553906
#> ES Sharpe (Rf=0.3%, p=95%):     0.1295014 0.1788256 0.2093343 0.06625013
#>                                       HAM5      HAM6 EDHEC LS EQ   SP500 TR
#> StdDev Sharpe (Rf=0.3%, p=95%): 0.03545540 0.3785371   0.3142695 0.12558293
#> VaR Sharpe (Rf=0.3%, p=95%):    0.02399862 0.3022965   0.2737607 0.07957460
#> ES Sharpe (Rf=0.3%, p=95%):     0.01664487 0.2308450   0.1855867 0.05760917
#>                                  US 10Y TR
#> StdDev Sharpe (Rf=0.3%, p=95%): 0.05684359
#> VaR Sharpe (Rf=0.3%, p=95%):    0.03741555
#> ES Sharpe (Rf=0.3%, p=95%):     0.02610548

SharpeRatio(edhec,Rf = .04/12)

#>                                 Convertible Arbitrage CTA Global
#> StdDev Sharpe (Rf=0.3%, p=95%):            0.15339751 0.12558803
#> VaR Sharpe (Rf=0.3%, p=95%):               0.09469804 0.09337063
#> ES Sharpe (Rf=0.3%, p=95%):                0.03089192 0.07366957
#>                                 Distressed Securities Emerging Markets
#> StdDev Sharpe (Rf=0.3%, p=95%):            0.25179740       0.12736677
#> VaR Sharpe (Rf=0.3%, p=95%):               0.16804482       0.07720662
#> ES Sharpe (Rf=0.3%, p=95%):                0.07589603       0.03841453
#>                                 Equity Market Neutral Event Driven
#> StdDev Sharpe (Rf=0.3%, p=95%):             0.2963971   0.23372885
#> VaR Sharpe (Rf=0.3%, p=95%):                0.2352556   0.15249824
#> ES Sharpe (Rf=0.3%, p=95%):                 0.0622208   0.07257669
#>                                 Fixed Income Arbitrage Global Macro
#> StdDev Sharpe (Rf=0.3%, p=95%):             0.06333844    0.2549431
#> VaR Sharpe (Rf=0.3%, p=95%):                0.03637036    0.2802548
#> ES Sharpe (Rf=0.3%, p=95%):                 0.01508086    0.2187777
#>                                 Long/Short Equity Merger Arbitrage
#> StdDev Sharpe (Rf=0.3%, p=95%):        0.19962891        0.3090483
#> VaR Sharpe (Rf=0.3%, p=95%):           0.14572983        0.2321344
#> ES Sharpe (Rf=0.3%, p=95%):            0.09399888        0.1037628
#>                                 Relative Value Short Selling Funds of Funds
#> StdDev Sharpe (Rf=0.3%, p=95%):     0.25525305   0.015024743     0.14194452
#> VaR Sharpe (Rf=0.3%, p=95%):        0.17482979   0.011139811     0.10328569
#> ES Sharpe (Rf=0.3%, p=95%):         0.06890142   0.009175123     0.05597178