wrapper to draw scatter plot with sensible defaults

Draws a scatter chart. This is another chart "primitive", since it only contains a set of sensible defaults.

chart.Scatter(x, y, reference.grid = TRUE, main = "Title", ylab = NULL,
  xlab = NULL, xlim = NULL, ylim = NULL, colorset = 1, symbolset = 1,
  element.color = "darkgray", cex.axis = 0.8, cex.legend = 0.8,
  cex.lab = 1, cex.main = 1, ...)

Arguments

x
data for the x axis, can take matrix,vector, or timeseries

y
data for the y axis, can take matrix,vector, or timeseries

reference.grid
if true, draws a grid aligned with the points on the x and y axes

main
set the chart title, same as in plot

ylab
set the y-axis label, as in plot

xlab
set the x-axis label, as in plot

xlim
set the x-axis limit, same as in plot

ylim
set the y-axis limit, same as in plot

colorset
color palette to use, set by default to rational choices

symbolset
from pch in plot, submit a set of symbols to be used in the same order as the data sets submitted

element.color
provides the color for drawing chart elements, such as the box lines, axis lines, etc. Default is "darkgray"

cex.axis
The magnification to be used for axis annotation relative to the current setting of 'cex', same as in plot.

cex.legend
The magnification to be used for sizing the legend relative to the current setting of 'cex'.

cex.lab
The magnification to be used for x- and y-axis labels relative to the current setting of 'cex'

cex.main
The magnification to be used for the main title relative to the current setting of 'cex'.

…
any other passthru parameters

Note

Most inputs are the same as "plot" and are principally included so that some sensible defaults could be set.

See also

plot

Examples

# NOT RUN {
  data(edhec)
  chart.Scatter(edhec[,1],edhec[,2])
# }