EDHEC-Risk Hedge Fund Style Indices

EDHEC composite hedge fund style index returns.

edhec

Details

EDHEC Data used in PerformanceAnalytics and related publications with the kind permission of the EDHEC Risk and Asset Management Research Center.

The 'edhec' data set included with PerformanceAnalytics will be periodically updated (typically annually) to include additional observations. If you intend to use this data set in automated tests, please be sure to subset your data like edhec[1:120,] to use the first ten years of observations.

From the EDHEC website: “The EDHEC Risk and Asset Management Research Centre plays a noted role in furthering applied financial research and systematically highlighting its practical uses. As part of its philosophy, the centre maintains a dialogue with professionals which benefits the industry as a whole. At the same time, its proprietary R&D provides sponsors with an edge over competition and joint ventures allow selected partners to develop new business opportunities. To further assist financial institutions and investors implement the latest research advances in order to meet the challenges of the changing asset management landscape, the centre has spawned two consultancies and an executive education arm. Clients of these derivative activities include many of the leading organisations throughout Europe.”

see http://www.edhec-risk.com/about_us

Format

CSV conformed into an xts object with monthly observations

Source

http://www.edhec-risk.com/indexes/pure_style

References

About EDHEC Alternative Indexes. December 16, 2003. EDHEC-Risk. http://www.edhec-risk.com/indexes/pure_style/about

Vaissie Mathieu. A Detailed Analysis of the Construction Methods and Management Principles of Hedge Fund Indices. October 2003. EDHEC. http://www.edhec-risk.com/site_edhecrisk/public/indexes/EDHEC_Publications/RISKReview1072705188065793513

Examples

data(edhec)

#preview the data
head(edhec)

#>            Convertible Arbitrage CTA Global Distressed Securities
#> 1997-01-31                0.0119     0.0393                0.0178
#> 1997-02-28                0.0123     0.0298                0.0122
#> 1997-03-31                0.0078    -0.0021               -0.0012
#> 1997-04-30                0.0086    -0.0170                0.0030
#> 1997-05-31                0.0156    -0.0015                0.0233
#> 1997-06-30                0.0212     0.0085                0.0217
#>            Emerging Markets Equity Market Neutral Event Driven
#> 1997-01-31           0.0791                0.0189       0.0213
#> 1997-02-28           0.0525                0.0101       0.0084
#> 1997-03-31          -0.0120                0.0016      -0.0023
#> 1997-04-30           0.0119                0.0119      -0.0005
#> 1997-05-31           0.0315                0.0189       0.0346
#> 1997-06-30           0.0581                0.0165       0.0258
#>            Fixed Income Arbitrage Global Macro Long/Short Equity
#> 1997-01-31                 0.0191       0.0573            0.0281
#> 1997-02-28                 0.0122       0.0175           -0.0006
#> 1997-03-31                 0.0109      -0.0119           -0.0084
#> 1997-04-30                 0.0130       0.0172            0.0084
#> 1997-05-31                 0.0118       0.0108            0.0394
#> 1997-06-30                 0.0108       0.0218            0.0223
#>            Merger Arbitrage Relative Value Short Selling Funds of Funds
#> 1997-01-31           0.0150         0.0180       -0.0166         0.0317
#> 1997-02-28           0.0034         0.0118        0.0426         0.0106
#> 1997-03-31           0.0060         0.0010        0.0778        -0.0077
#> 1997-04-30          -0.0001         0.0122       -0.0129         0.0009
#> 1997-05-31           0.0197         0.0173       -0.0737         0.0275
#> 1997-06-30           0.0231         0.0198       -0.0065         0.0225

#summary period statistics
summary(edhec)

#>      Index            Convertible Arbitrage   CTA Global       
#>  Min.   :1997-01-31   Min.   :-0.123700     Min.   :-0.054300  
#>  1st Qu.:2000-03-23   1st Qu.: 0.000925     1st Qu.:-0.011500  
#>  Median :2003-05-15   Median : 0.009200     Median : 0.005250  
#>  Mean   :2003-05-16   Mean   : 0.006409     Mean   : 0.006489  
#>  3rd Qu.:2006-07-07   3rd Qu.: 0.014600     3rd Qu.: 0.022675  
#>  Max.   :2009-08-31   Max.   : 0.061100     Max.   : 0.069100  
#>  Distressed Securities Emerging Markets    Equity Market Neutral
#>  Min.   :-0.083600     Min.   :-0.192200   Min.   :-0.058700    
#>  1st Qu.:-0.000175     1st Qu.:-0.011400   1st Qu.: 0.002400    
#>  Median : 0.009700     Median : 0.013650   Median : 0.006300    
#>  Mean   : 0.007953     Mean   : 0.008246   Mean   : 0.006003    
#>  3rd Qu.: 0.018225     3rd Qu.: 0.028875   3rd Qu.: 0.010125    
#>  Max.   : 0.050400     Max.   : 0.123000   Max.   : 0.025300    
#>   Event Driven       Fixed Income Arbitrage  Global Macro      
#>  Min.   :-0.088600   Min.   :-0.086700      Min.   :-0.031300  
#>  1st Qu.:-0.000025   1st Qu.: 0.002275      1st Qu.:-0.003600  
#>  Median : 0.010150   Median : 0.006000      Median : 0.006200  
#>  Mean   : 0.007622   Mean   : 0.004231      Mean   : 0.007672  
#>  3rd Qu.: 0.018650   3rd Qu.: 0.009950      3rd Qu.: 0.016450  
#>  Max.   : 0.044200   Max.   : 0.036500      Max.   : 0.073800  
#>  Long/Short Equity  Merger Arbitrage    Relative Value      Short Selling      
#>  Min.   :-0.06750   Min.   :-0.054400   Min.   :-0.069200   Min.   :-0.134000  
#>  1st Qu.:-0.00370   1st Qu.: 0.001775   1st Qu.: 0.001850   1st Qu.:-0.025450  
#>  Median : 0.01015   Median : 0.007700   Median : 0.008450   Median :-0.000200  
#>  Mean   : 0.00776   Mean   : 0.006785   Mean   : 0.006701   Mean   : 0.004161  
#>  3rd Qu.: 0.02145   3rd Qu.: 0.013725   3rd Qu.: 0.014500   3rd Qu.: 0.035925  
#>  Max.   : 0.07450   Max.   : 0.027200   Max.   : 0.039200   Max.   : 0.246300  
#>  Funds of Funds     
#>  Min.   :-0.061800  
#>  1st Qu.:-0.003325  
#>  Median : 0.006800  
#>  Mean   : 0.005918  
#>  3rd Qu.: 0.015225  
#>  Max.   : 0.066600  

#cumulative index returns
tail(cumprod(1+edhec),1)

#>            Convertible Arbitrage CTA Global Distressed Securities
#> 2009-08-31              2.559585   2.550288              3.249667
#>            Emerging Markets Equity Market Neutral Event Driven
#> 2009-08-31         3.106494              2.468271     3.091342
#>            Fixed Income Arbitrage Global Macro Long/Short Equity
#> 2009-08-31               1.870396        3.128          3.120998
#>            Merger Arbitrage Relative Value Short Selling Funds of Funds
#> 2009-08-31         2.768785       2.723671      1.502321        2.39178