Hypothetical Alternative Asset Manager and Benchmark Data

A xts object that contains columns of monthly returns for six hypothetical asset managers (HAM1 through HAM6), the EDHEC Long-Short Equity hedge fund index, the S\&P 500 total returns, and total return series for the US Treasury 10-year bond and 3-month bill. Monthly returns for all series end in December 2006 and begin at different periods starting from January 1996.

Note that all the EDHEC indices are available in edhec.

managers

Details

Please note that the `managers' data set included with PerformanceAnalytics will be periodically updated with new managers and information. If you intend to use this data set in automated tests, please be sure to subset your data like managers[1:120,1:6] to use the first ten years of observations on HAM1-HAM6.

Format

CSV conformed into an xts object with monthly observations

Examples

data(managers)

#preview the data
head(managers)

#>               HAM1 HAM2    HAM3    HAM4 HAM5 HAM6 EDHEC LS EQ SP500 TR
#> 1996-01-31  0.0074   NA  0.0349  0.0222   NA   NA          NA   0.0340
#> 1996-02-29  0.0193   NA  0.0351  0.0195   NA   NA          NA   0.0093
#> 1996-03-31  0.0155   NA  0.0258 -0.0098   NA   NA          NA   0.0096
#> 1996-04-30 -0.0091   NA  0.0449  0.0236   NA   NA          NA   0.0147
#> 1996-05-31  0.0076   NA  0.0353  0.0028   NA   NA          NA   0.0258
#> 1996-06-30 -0.0039   NA -0.0303 -0.0019   NA   NA          NA   0.0038
#>            US 10Y TR US 3m TR
#> 1996-01-31   0.00380  0.00456
#> 1996-02-29  -0.03532  0.00398
#> 1996-03-31  -0.01057  0.00371
#> 1996-04-30  -0.01739  0.00428
#> 1996-05-31  -0.00543  0.00443
#> 1996-06-30   0.01507  0.00412

#summary period statistics
summary(managers)

#>      Index                 HAM1                HAM2         
#>  Min.   :1996-01-31   Min.   :-0.094400   Min.   :-0.03710  
#>  1st Qu.:1998-10-23   1st Qu.:-0.000025   1st Qu.:-0.00980  
#>  Median :2001-07-15   Median : 0.011150   Median : 0.00820  
#>  Mean   :2001-07-15   Mean   : 0.011123   Mean   : 0.01414  
#>  3rd Qu.:2004-04-07   3rd Qu.: 0.024850   3rd Qu.: 0.02520  
#>  Max.   :2006-12-31   Max.   : 0.069200   Max.   : 0.15560  
#>                                           NA's   :7         
#>       HAM3                HAM4               HAM5               HAM6         
#>  Min.   :-0.071800   Min.   :-0.17590   Min.   :-0.13200   Min.   :-0.04040  
#>  1st Qu.:-0.005375   1st Qu.:-0.01985   1st Qu.:-0.01640   1st Qu.:-0.00158  
#>  Median : 0.010200   Median : 0.01375   Median : 0.00380   Median : 0.01285  
#>  Mean   : 0.012447   Mean   : 0.01102   Mean   : 0.00409   Mean   : 0.01105  
#>  3rd Qu.: 0.031375   3rd Qu.: 0.04600   3rd Qu.: 0.03090   3rd Qu.: 0.02548  
#>  Max.   : 0.179600   Max.   : 0.15080   Max.   : 0.17470   Max.   : 0.05830  
#>                                         NA's   :55         NA's   :68        
#>   EDHEC LS EQ           SP500 TR           US 10Y TR            US 3m TR       
#>  Min.   :-0.055200   Min.   :-0.144600   Min.   :-0.070920   Min.   :0.000660  
#>  1st Qu.:-0.003175   1st Qu.:-0.017327   1st Qu.:-0.008515   1st Qu.:0.001570  
#>  Median : 0.011000   Median : 0.010950   Median : 0.004425   Median :0.003845  
#>  Mean   : 0.009545   Mean   : 0.008665   Mean   : 0.004385   Mean   :0.003226  
#>  3rd Qu.: 0.021450   3rd Qu.: 0.038025   3rd Qu.: 0.016660   3rd Qu.:0.004395  
#>  Max.   : 0.074500   Max.   : 0.097800   Max.   : 0.050550   Max.   :0.006580  
#>  NA's   :12                                                                    

#cumulative returns
tail(cumprod(1+managers),1)

#>                HAM1 HAM2     HAM3    HAM4 HAM5 HAM6 EDHEC LS EQ SP500 TR
#> 2006-12-31 4.126671   NA 4.706732 3.52944   NA   NA          NA 2.761619
#>            US 10Y TR US 3m TR
#> 2006-12-31  1.734037 1.529681