calculate attributes relative to the mean of the observation series given, including geometric, stderr, LCL and UCL

mean.geometric 	 geometric mean

mean.stderr 	 standard error of the mean (S.E. mean)

mean.LCL 	 lower confidence level (LCL) of the mean

mean.UCL 	 upper confidence level (UCL) of the mean

# S3 method for geometric
mean(x, ...)

# S3 method for stderr
mean(x, ...)

# S3 method for LCL
mean(x, ci = 0.95, ...)

# S3 method for UCL
mean(x, ci = 0.95, ...)

Arguments

x
a vector, matrix, data frame, or time series to calculate the modified mean statistic over

…
any other passthru parameters

ci
the confidence interval to use

See also

sd mean

Examples

data(edhec)
mean.geometric(edhec[,"Funds of Funds"])

#>                Funds of Funds
#> Geometric Mean    0.005753581

mean.stderr(edhec[,"Funds of Funds"])

#>                Funds of Funds
#> Standard Error    0.001477185

mean.UCL(edhec[,"Funds of Funds"])

#>                        Funds of Funds
#> Upper Confidence Level    0.008837041

mean.LCL(edhec[,"Funds of Funds"])

#>                        Funds of Funds
#> Lower Confidence Level    0.002999801