Bacon(2008) Data

A xts object that contains columns of monthly returns for an example of portfolio and its benchmark

portfolio_bacon

Format

CSV conformed into an xts object with monthly observations

Examples

data(portfolio_bacon)

#preview the data
head(portfolio_bacon)

#> Warning: index class is Date, which does not support timezones.
#> Expected 'UTC' timezone, but indexTZ is ''

#>            portfolio.monthly.return.... benchmark.return....
#> 2000-01-30                        0.003                0.002
#> 2000-02-27                        0.026                0.025
#> 2000-03-30                        0.011                0.018
#> 2000-04-29                       -0.010               -0.011
#> 2000-05-30                        0.015                0.014
#> 2000-06-29                        0.025                0.018

#summary period statistics
summary(portfolio_bacon)

#>      Index            portfolio.monthly.return.... benchmark.return....
#>  Min.   :2000-01-30   Min.   :-0.0650              Min.   :-0.06700    
#>  1st Qu.:2000-07-22   1st Qu.:-0.0110              1st Qu.:-0.00725    
#>  Median :2001-01-14   Median : 0.0130              Median : 0.01450    
#>  Mean   :2001-01-13   Mean   : 0.0090              Mean   : 0.01004    
#>  3rd Qu.:2001-07-06   3rd Qu.: 0.0295              3rd Qu.: 0.02850    
#>  Max.   :2001-12-30   Max.   : 0.0810              Max.   : 0.08300    

#cumulative returns
tail(cumprod(1+portfolio_bacon),1)

#> Warning: index class is Date, which does not support timezones.
#> Expected 'UTC' timezone, but indexTZ is ''

#>            portfolio.monthly.return.... benchmark.return....
#> 2001-12-30                     1.218106             1.249887