Distributions Summary: Statistics and Stylized Facts

Table of Monthly standard deviation, Skewness, Sample standard deviation, Kurtosis, Excess kurtosis, Sample Skweness and Sample excess kurtosis

table.Distributions(R, scale = NA, digits = 4)

Arguments

R
an xts, vector, matrix, data frame, timeSeries or zoo object of asset returns

scale
number of periods in a year (daily scale = 252, monthly scale = 12, quarterly scale = 4)

digits
number of digits to round results to

References

Carl Bacon, Practical portfolio performance measurement and attribution, second edition 2008 p.87

See also

StdDev.annualized skewness kurtosis

Examples

data(managers)
table.Distributions(managers[,1:8])

#>                           HAM1   HAM2   HAM3    HAM4   HAM5    HAM6 EDHEC LS EQ
#> Monthly Std Dev         0.0256 0.0367 0.0365  0.0532 0.0457  0.0238      0.0205
#> Skewness               -0.6588 1.4580 0.7908 -0.4311 0.0738 -0.2800      0.0177
#> Kurtosis                5.3616 5.3794 5.6829  3.8632 5.3143  2.6511      3.9105
#> Excess kurtosis         2.3616 2.3794 2.6829  0.8632 2.3143 -0.3489      0.9105
#> Sample skewness        -0.6741 1.4937 0.8091 -0.4410 0.0768 -0.2936      0.0182
#> Sample excess kurtosis  2.5004 2.5270 2.8343  0.9437 2.5541 -0.2778      1.0013
#>                        SP500 TR
#> Monthly Std Dev          0.0433
#> Skewness                -0.5531
#> Kurtosis                 3.5598
#> Excess kurtosis          0.5598
#> Sample skewness         -0.5659
#> Sample excess kurtosis   0.6285

require("Hmisc")

#> Loading required package: Hmisc

#> Warning: there is no package called ‘Hmisc’

result = t(table.Distributions(managers[,1:8]))

textplot(format.df(result, na.blank=TRUE, numeric.dollar=FALSE, cdec=c(3,3,1)),
rmar = 0.8, cmar = 2,  max.cex=.9, halign = "center", valign = "top",
row.valign="center", wrap.rownames=20, wrap.colnames=10,
col.rownames=c("red", rep("darkgray",5), rep("orange",2)), mar = c(0,0,3,0)+0.1)

#> Error in format.df(result, na.blank = TRUE, numeric.dollar = FALSE, cdec = c(3,     3, 1)): could not find function "format.df"

title(main="Portfolio Distributions statistics")

#> Error in title(main = "Portfolio Distributions statistics"): plot.new has not been called yet