Selected Portfolio Weights Data

An xts object that contains columns of monthly weights for a subset of the EDHEC hedge fund indexes that demonstrate rebalancing portfolios through time.

Note that all the EDHEC indices are available in edhec.

managers

Details

A relatively random weights file used for charting examples.

Format

CSV conformed into an xts object with monthly observations

Examples

data(weights)

#preview the data
head(weights)

#>            Convertible Arbitrage CTA Global Distressed Securities
#> 2000-01-01            0.02500000 0.14601749             0.0250000
#> 2001-01-01            0.15785710 0.19577551             0.0250000
#> 2002-01-01            0.24431295 0.02500000             0.0250000
#> 2003-01-01            0.21955470 0.06590151             0.0250000
#> 2004-01-01            0.09780634 0.02552822             0.1050766
#> 2005-01-01            0.02500000 0.02500000             0.2445763
#>            Emerging Markets Equity Market Neutral Event Driven
#> 2000-01-01            0.025             0.3500000        0.025
#> 2001-01-01            0.025             0.3500000        0.025
#> 2002-01-01            0.025             0.3500000        0.025
#> 2003-01-01            0.025             0.2817930        0.025
#> 2004-01-01            0.025             0.3500000        0.025
#> 2005-01-01            0.025             0.2054237        0.025
#>            Fixed Income Arbitrage Global Macro Long/Short Equity
#> 2000-01-01              0.0250000        0.025             0.025
#> 2001-01-01              0.0250000        0.025             0.025
#> 2002-01-01              0.2056871        0.025             0.025
#> 2003-01-01              0.2577508        0.025             0.025
#> 2004-01-01              0.2715888        0.025             0.025
#> 2005-01-01              0.3500000        0.025             0.025
#>            Merger Arbitrage Relative Value
#> 2000-01-01       0.07146246      0.2575201
#> 2001-01-01       0.12136740      0.0250000
#> 2002-01-01       0.02500000      0.0250000
#> 2003-01-01       0.02500000      0.0250000
#> 2004-01-01       0.02500000      0.0250000
#> 2005-01-01       0.02500000      0.0250000